# src/main/webapp/swagger/index.yaml
swagger: '2.0'
info:
  description: JAVA后台管理系统，企业开发首选脚手架。
  version: 1.0.0
  title: SpringAdmin
basePath: /SpringAdmin
schemes:
  - http

#定义接口数据
paths:
...
  /sys/dept/list:
    get:
      tags:
        - 部门管理
      summary: 查询部门列表
      produces:
        - application/json
      parameters:
        - name: sortName
          description: 排序字段
          in: query
          type: string
          required: true
        - name: sortOrder
          description: 排序方式，如：asc、desc
          in: query
          type: string
          required: true
      responses:
        '200':
          description: 部门列表
          schema:
            $ref: '#/definitions/SysDeptInfo'
  /sys/dept/tree:
    post:
      tags:
        - 部门管理
      summary: 选择部门
      description: 添加、修改部门的时候，选择上级部门接口
      produces:
        - application/json
      parameters:
        - name: depId
          description: 部门ID
          in: body
          type: string
      responses:
        '200':
          description: 部门列表
          schema:
            $ref: '#/definitions/SysDeptInfo'
  /sys/dept/save:
    post:
      tags:
        - 部门管理
      summary: 新增部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/dept/update:
    post:
      tags:
        - 部门管理
      summary: 修改部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysDeptInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/dept/delete:
    post:
      tags:
        - 部门管理
      summary: 删除部门
      produces:
        - application/json
      parameters:
        - name: body
          description: 部门ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
...
  /sys/res/list:
    get:
      tags:
        - 资源管理
      summary: 查询资源列表
      produces:
        - application/json
      parameters:
        - name: sortName
          description: 排序字段
          in: query
          type: string
          required: true
        - name: sortOrder
          description: 排序方式，如：asc、desc
          in: query
          type: string
          required: true
      responses:
        '200':
          description: 返回资源列表
          schema:
            $ref: '#/definitions/SysResInfo'
...
  /sys/res/save:
    post:
      tags:
        - 资源管理
      summary: 添加资源
      produces:
        - application/json
      parameters:
        - name: body
          description: 资源对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysResInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/res/update:
    post:
      tags:
        - 资源管理
      summary: 修改资源
      produces:
        - application/json
      parameters:
        - name: body
          description: 资源对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysResInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/res/delete:
    post:
      tags:
        - 资源管理
      summary: 删除资源
      produces:
        - application/json
      parameters:
        - name: body
          description: 资源ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
...
  /sys/job/list:
    get:
      tags:
        - 定时任务
      summary: 定时任务列表
      produces:
        - application/json
      parameters:
        - name: pageSize
          description: 分页条数
          in: query
          type: integer
          required: true
        - name: pageNumber
          description: 分页页码
          in: query
          type: integer
          required: true
        - name: sortName
          description: 排序字段
          in: query
          type: string
          required: true
        - name: sortOrder
          description: 排序方式，如：asc、desc
          in: query
          type: string
          required: true
        - name: beanName
          description: bean名称
          in: query
          type: string
        - name: status
          description: 状态
          in: query
          type: string
      responses:
        '200':
          description: 定时任务列表
          schema:
            $ref: '#/definitions/SysJobInfo'
  /sys/job/save:
    post:
      tags:
        - 定时任务
      summary: 新增定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysJobInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/job/update:
    post:
      tags:
        - 定时任务
      summary: 修改定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务对象
          in: body
          type: string
          schema:
            $ref: '#/definitions/SysJobInfo'
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/job/delete:
    post:
      tags:
        - 定时任务
      summary: 删除定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/job/run:
    post:
      tags:
        - 定时任务
      summary: 立即执行任务
      produces:
        - application/json
      parameters:
        - name: ids
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/job/pause:
    post:
      tags:
        - 定时任务
      summary: 暂停定时任务
      produces:
        - application/json
      parameters:
        - name: ids
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'
  /sys/job/resume:
    post:
      tags:
        - 定时任务
      summary: 恢复定时任务
      produces:
        - application/json
      parameters:
        - name: body
          description: 定时任务ID列表
          in: body
          type: array
          items:
            type: integer
            format: int64
            default: 0
          required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/Result'

#定义数据模型
definitions:
  Result:
      type: object
      properties:
        code:
          description: 状态码 1：成功  0：失败
          type: integer
          format: int32
        msg:
          description: 失败原因
          type: string
...
  SysDeptInfo:
    type: object
    properties:
      id:
        description: 部门ID
        type: integer
        format: int64
      deptName:
        description: 部门名称
        type: string
      deptDesc:
        description: 部门描述
        type: string
      parentId:
        description: 上级部门ID，一级部门为0
        type: integer
        format: int64
      parentName:
        description: 上级部门名称
        type: string
      orderNo:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
...
  SysResInfo:
    type: object
    properties:
      id:
        description: 资源ID
        type: integer
        format: int64
      resName:
        description: 资源名称
        type: string
      resDesc:
        description: 资源描述
        type: string
      parentId:
        description: 父资源ID，一级资源为0
        type: integer
        format: int64
      parentName:
        description: 父资源名称
        type: string
      resUrl:
        description: 资源URL
        type: string
      resPerms:
        description: 资源权限标识
        type: string
      resType:
        description: 资源类型     0：目录   1：资源   2：按钮
        type: integer
        format: int32
      resIcon:
        description: 资源图标
        type: string
      orderNo:
        description: 排序
        type: integer
        format: int32
      open:
        description: 是否展开   true：展开   false：不展开
        type: boolean
        format: int32
      createTime:
        description: 创建时间
        type: string
        format: date-time
      updateTime:
        description: 更新时间
        type: string
        format: date-time
...
  SysJobInfo:
    type: object
    properties:
      id:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
      updateTime:
        description: 更新时间
        type: string
        format: date-time